test/web-platform-tests/to-upstream/svg/SVGElement-ownerSVGElement.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>SVGElement.prototype.ownerSVGElement</title>
<link rel=help href="https://svgwg.org/svg2-draft/types.html#__svg__SVGElement__ownerSVGElement">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<body>
  <svg id="test1">
    <svg id="test2">
      <symbol id="test3">
        <svg id="test4"></svg>
      </symbol>
    </svg>
  </svg>
</body>

<script>
"use strict";

test(() => {
  const svg1 = document.getElementById("test1");
  const svg2 = document.getElementById("test2");
  const svg3 = document.getElementById("test4");

  assert_equals(svg1.ownerSVGElement, null);
  assert_equals(svg2.ownerSVGElement.id, "test1");
  assert_equals(svg3.ownerSVGElement.id, "test2");
});
</script>
